Add TomlStreamSerializer
Introduced the new TOML object visitor. This new class allows to remove the runtime type checking. Instead
the double dispatch pattern is used, so that the compiler resolves the correct type during compile time.
This allows to remove the WriteTo stuff from the TOML objects and extract the serialization logic
into a single class. Also we can get rid of the IsArray ... helper stuff that makes the objects themself cleaner.
Also some of the useless overloads for Get/Write were removed.