Implicit TOML table arrays working
With this change all CocoR working tests are working again. Even some
other tests I didn't manage to get to work with it, are now green :-).
With this I can finally start to implement some missing features and
start to extend the test set (I already know some corner cases are
still broken and a performance profiling test suite should be
established as soon as possible). So much to do so little time ... :-)