Better fluent configuration API
The new API is a lot simpler internally and in extenal usage, as it
removes the combination specifiers And/As that only bloated the
configuration.

It start to establish a concept of how Nett will handle configuration
nesting (Configure(), Apply()).

The Configure() call allows to jump back to root level, while Apply()
will go up one level in the configuration. This new config system is
much more like e.g. EntityFrameworks (6.1) config system, but in
addition still allows you to change nested config levels.

In contrast in EF once you are on the property config level you won't
be able to get back to the entity config level. One needs to start a
second entity config to do so. In Nett this should be possible, altough
it makes the API more complicated.