Lambda expression based config API
This removes the need for Apply() and GetConfig() completely.
Also makes the creation of a config an atomic operation when the
builder API is used. Either the whole config is applied or nothing at
all.

It is easier to distinguish the different config levels because of the
braces.

Altough this syntax needs a little time to get used to it is a really
nice, extensible and clean way to do complex configuration stuff in
code.