Separate test project into interal/API test projects
+ Also updates nuget packages
   + Fluent Assertions
   + XUnit

This separation is the starting point for the new upcoming config
API. The new API requires that we can dinstinguish between internal
tests (as currently public API surface will be made internal) and
public API tests that check that the API wasn't broke accidentally.

Some of this API modifications are already in this commit to get
the stuff to compile again without weird work arounds.