Change conversion config API to use CLR types for simple TOML types
The user doesn't need to specify the TOML object construction statement
in the config he creates. Instead in the configuraiton the equivalent
CLR type gets returned. The underlying TOML system will create the TOML
object.

This is preparation for making parts of the TOML API internal to make
it easier to use and ensure the user isn't able to produce weird data
states.