Add MetaDataStore concept
This is a huge change in the architecture but will enable a a lot of
advanced usage scenarios. The MetaDataStore allows to attach Nett
iternal information to TOML data structures.

Currently only the config that was used during reads/creates of tables
is attached. But by attaching it, it is much easier for the consumer of
the API to extract data from generic structures using the same config.

The nice thing is, that the managemnt of this metadata is completely
hidden on the consumer side (the only thing the API consumer will
notice, that it is not possible to create TOML types directly).

The public API is still not finished with this commit. This commit gets
everything to compile again and makes sure the existing test are green.
Further commits will finish this new stuff.