Writing back into specific scopes works
This introduces a quite hefty API change for the typed config. But
currently I don't see any way to get around it.

Saving into specific sources was not possible with the old mechanism of
transforming the TOML table to CLR object and operating on that. This
is because in the update cases it was impossible to track what was
changed and only save that changes back to the TOML file.

This is causes by the TOML / CLR mismatch... While in TOML table rows
can be missing, a property of an CLR object always has to be present,
this causes a lot of issues for the write back case.

The change is form config.Set(c => c.X = 1) to config.Set(c => c.X, 1)
a form that still looks quite good for all the nice fluff provided by
the background auto magic.

Another positive aspect is, that it should be more faster as mapping
from TomlTable to CLR type is not needed anymore.