Nett: TOML with value not in same line as key will produce parse error
Was done by a parser change. The parser now doesn't ignore
the newlines anymore.

That means all language fragments where newlines are allowed
need to handle them in their parser productions.

So the productions get a little bit more complicated, but the
good thing is that a hack could be removed.