BareKey matcher checks that token is finished before token is returned
This will provide better error messages for cases where some other token
directly followed up to a bare key.

E.g.
X = 1; #1 would be classified as bare key in old system, now as a
unknown token with value '1;'

This more errornous behavior is caused because bare bare keys
are a little bit more hard to classify as they do not have
a special symbol and also things like 1234 are allowed as bare
keys.