Fix #23: Write key back to file with same type
If a read key was a bare, basic, literal key durin the read operation,
it will also get written back to the file with the same type again.

The key is not a simple string any more in the internal data structures
but a complex type instead. This type 'remembers' what key type it was
during the read operation of a file.

This allows that the writer operation of such an table to write the
key with the same type again.

Additionally the following things were done to enable this change

1. Toml tables do not allow to add other TomlTables as sub tables
anymore. This was allowed by accident and enabled API users to
create broken TomlTable Graphs. Instead a new Add(string) method
that adds an empty sub table that can later be modified was
added instead.

2. The internal Add method was renamed to AddRow to reduce the
possibility of OverloadResolution issues.