Disallow properties to be treated as inline tables
- Renames the attribute to be consistent with the fluent API

Although it may seem to reduce flexibility of the library by doing so
in reality this commit removes a design issue of Nett.

The possibility to selectively select properties that get treated as
inline table while others are not even when the have the same type
encouraged API users to produce inconsistent configs.

If a row's value should get written as an inline table or not is
determined by its type, not where the type is used.

This change also simplifies the get table type operation and makes it
more consistent among its usages (null param not needed anymore).

Related: #19