Add test showing TPath design flaws
Current TPath is not able to create TOML objects from itself (for non
leaf objects). This causes issues when Coma tries to save settings that
are not contained in that config yet.

These two test cases are the first to show that flaw and are the
stating point to eliminate this design flaw.

This commit is expected to cause a failing build.