Parse arrays with comments correctly
The old Nett impl was and ABNF compatible solution. But the old ABNF
specification for TOML was not compatible with the latest TOML
specification.

The new implementation allows comments in a lot more places. But, this
implementation does not preserve the comments in a read/write cycle
yet. This will take some more work and some redesign.

Comments processing overall is still a somewhat shaky process. This
should be improved in the future.

Related: #30