Doc: Fix documentation generatino for base projects
Documentation generation was broken from the switch to multi target
framework projects.

https://github.com/dotnet/docfx/issues/2304

For now we activate the Net40 generation so that the standard projects
generate documentation again. This will break the documentation generation
for Nett.AspNet for now.

For all documentation generation to work correctly again
I will wait until
https://github.com/dotnet/docfx/issues/2322
is implementd.