Doc: Use DocFX as doucmentation system
- Split README.md
- Create a dedicated DocFX documentation site
- Add various sections
- Try to do first doc clean up

This commits purpose is not to create a more
completed docu. This commit focuses
on improving the technical basis for the
documentationm, so that in future it can be
improved and extended easier.

It will be a long road until the doucumentation is
really usefull - especially with my godlike *ehmehm*
documentation skills.