AspNet: Added a way to configure TomlConfigurationProvider to load a case insensitive data dictionary.
Implements issue #81 by adding an overload to ProviderDictionaryConverter.ToProviderDictionary that accepts a StringComparer for the created dictionary.

TomlConfigurationProvider now calls this overload with a StringComparer.OrdinalIgnore case value if the added CaseInsensitiveKeys property is set to true.

The CaseInsensitiveKeys property on the provider can be set by modifying the identically named property on TomlConfigurationSource.

In turn, CaseInsensitiveKeys on the source can be set by calling an AddTomlFile overload that has a configureSource parameter.

Also replaced the provider and path checks in the main AddTomlFile method with a call to ResolveFileProvider which does the same thing.