Mapping concrete types into generic base types
User can now map TOML objects into specific base types via
specifying a concrete activator. A consequence is that the
ambiguous API MapTableKey is not needed anymore. This
is a good thing but will introduce a breaking change with
the next release.

While this commit makes it work and all test are green, I have
the feeling not all scenarios will work as e.g. the ToDictionary
test are not good enough to catch all issues here.

In general the conversion from TOML object graph to a CLR
object graph has a lot of technical debt. E.g. conversion is
partially handled by converter and partially handled by the
TOML objects themselves. The visitor pattern there is a pain.
Future work will be needed to improve the code base in this
space.

Another consequence of this inconsistencies is, that the
CreateInstanceContext KeyChain property most of the time
is not filled with correct data.

Related #90